+++ lattpy/tests/test_data.py
@@ -115,21 +115,53 @@ def test_datamap():
     latt.build(5)
 
     dmap = latt.data.map()
-    res = [True, True, True, True, True, True, False, False, False,
-           False, False, False, False, False, False, False]
+    res = [
+        True,
+        True,
+        True,
+        True,
+        True,
+        True,
+        False,
+        False,
+        False,
+        False,
+        False,
+        False,
+        False,
+        False,
+        False,
+        False,
+    ]
     assert dmap.size == len(res)
     assert_array_equal(dmap.onsite(0), res)
 
-    res = [False, False, False, False, False, False, True, True, True,
-           True, True, True, True, True, True, True]
+    res = [
+        False,
+        False,
+        False,
+        False,
+        False,
+        False,
+        True,
+        True,
+        True,
+        True,
+        True,
+        True,
+        True,
+        True,
+        True,
+        True,
+    ]
     assert_array_equal(dmap.hopping(0), res)
 
     data = np.zeros(dmap.size)
     dmap.fill(data, hop=1, eps=2)
 
     res = np.zeros(dmap.size)
-    res[:latt.num_sites] = 2.0
-    res[latt.num_sites:] = 1.0
+    res[: latt.num_sites] = 2.0
+    res[latt.num_sites :] = 1.0
     assert_array_equal(data, res)
 
 
@@ -138,12 +170,62 @@ def test_site_mask():
     latt.build((4, 4))
     data = latt.data
 
-    expected = [0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1,
-                1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]
+    expected = [
+        0,
+        0,
+        0,
+        0,
+        0,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+        1,
+    ]
     assert_array_equal(data.site_mask([1, 0]), expected)
 
-    expected = [0, 1, 1, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1,
-                1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 1]
+    expected = [
+        0,
+        1,
+        1,
+        1,
+        1,
+        0,
+        1,
+        1,
+        1,
+        1,
+        0,
+        1,
+        1,
+        1,
+        1,
+        0,
+        1,
+        1,
+        1,
+        1,
+        0,
+        1,
+        1,
+        1,
+        1,
+    ]
     assert_array_equal(data.site_mask([0, 1]), expected)